requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fonttools==4.53.1
greenlet==3.0.3
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.11.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.2.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13